move options code
