+++ app.go
@@ -40,6 +40,25 @@ func NewWithMackerelClient(client MackerelClient, cfg *Config) (*App, error) {
 	return app, nil
 }
 
+type Options struct {
+	dryRun   bool
+	backfill int
+}
+
+//DryRunOption is an option to output the calculated error budget as standard without posting it to Mackerel.
+func DryRunOption(dryRun bool) func(*Options) {
+	return func(opt *Options) {
+		opt.dryRun = dryRun
+	}
+}
+
+//BackfillOption specifies how many points of data to calculate retroactively from the current time.
+func BackfillOption(count int) func(*Options) {
+	return func(opt *Options) {
+		opt.backfill = count
+	}
+}
+
 //Run performs the calculation of the error bar calculation
 func (app *App) Run(ctx context.Context, optFns ...func(*Options)) error {
 	orgName, err := app.repo.GetOrgName(ctx)
